fix: socio register and list

+++ app/src/main/java/com/odvsystem/sportcenter/model/Socio.kt
@@ -2,7 +2,7 @@ package com.odvsystem.sportcenter.model
 
 data class Socio (
     val nrosocio: Int,
-    val idusuario: Long,
+    val idusuario: Int,
     val estadohabilitacion: String,
     val cuotamensual: Double,
     val carneteentregado: Int,
